+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * :feature:`18` Use JSONSchema validation for actor keyword datamodel. The datamodel of the actor has been updated. The schema should work for all CLU actors, including ``LegacyActor`` although in this case the keywords will be flattened into a list.
 * :support:`18` Improve notifications during an exposure.
+* :feature:`-` `.CardGroup` now accepts string items that are evaluated to default cards.
 
 * :release:`0.4.1 <2021-02-13>`
 * :feature:`13` Add a `.WCSCards` macro that expands into full WCS header information.